split MCD class to its own file
